requirements.txt
torch
torchvision
torchxrayvision
streamlit
opencv-python
scikit-learn
matplotlib
numpy
Pillow
tqdm
